Graph: store samples as int32 deci-units; use Tiny5 for current value
The ring buffer stored samples as int16 x10 fixed-point, which saturated
at INT16_MAX/10 = 3276.7 — so high-magnitude tokens like power in watts
(e.g. [HA:sensor.ground_power] at 4357.2) were silently clamped and the
graph flat-lined at 3276.7 even though the message/token path showed the
real value. Widen the ring to int32 (covers ~±214M real units) with a
distinct gap sentinel GRAPH_SAMPLE_UNSET, kept separate from the int16
config sentinel GRAPH_VAL_UNSET. Config threshold fields stay int16.

Also switch the graph's current-value readout to the compact Tiny5 font
to match the axis labels. Host test updated (int32 out buffers, x10
expectations, and a regression case for the 4357.2 -> 43572 store).

+++ main/include/f-graph.h
@@ -26,7 +26,7 @@
  */
 typedef struct
 {
-  int16_t v[GRAPH_MAX_POINTS]; // samples; GRAPH_VAL_UNSET = gap / no sample
+  int32_t v[GRAPH_MAX_POINTS]; // samples (value x10); GRAPH_SAMPLE_UNSET = gap / no sample
   char token[GRAPH_TOKEN_LEN]; // token currently sampled, e.g. "[temp]"
   uint8_t count;               // valid samples (0..cap)
   uint8_t cap;                 // active capacity = cfg.points
@@ -48,7 +48,7 @@ void graph_configure(const screen_graph_cfg_t *cfg, bool widget_enabled); // on 
 
 // Snapshot up to max chronological samples (oldest first) into out_v. Returns
 // the number copied; optionally returns the anchor time and interval.
-int graph_snapshot(int16_t *out_v, int max, time_t *last_sample, uint16_t *interval_min);
+int graph_snapshot(int32_t *out_v, int max, time_t *last_sample, uint16_t *interval_min);
 
 bool graph_is_active(void);
